+++ src/main/java/org/com/itpple/spot/server/config/SecurityConfig.java
@@ -2,11 +2,11 @@ package org.com.itpple.spot.server.config;
 
 
 import lombok.RequiredArgsConstructor;
-import org.com.itpple.spot.server.common.jwt.JwtAccessDeniedHandler;
-import org.com.itpple.spot.server.common.jwt.JwtAuthenticationEntryPoint;
-import org.com.itpple.spot.server.common.jwt.JwtFilter;
-import org.com.itpple.spot.server.common.jwt.TokenProvider;
-import org.com.itpple.spot.server.common.jwt.UserDetailsCustomService;
+import org.com.itpple.spot.server.common.auth.jwt.JwtAccessDeniedHandler;
+import org.com.itpple.spot.server.common.auth.jwt.JwtAuthenticationEntryPoint;
+import org.com.itpple.spot.server.common.auth.jwt.JwtFilter;
+import org.com.itpple.spot.server.common.auth.jwt.TokenProvider;
+import org.com.itpple.spot.server.common.auth.userDetails.UserDetailsCustomService;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
