+++ src/main/java/org/com/itpple/spot/server/controller/PotController.java
@@ -3,7 +3,7 @@ package org.com.itpple.spot.server.controller;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.com.itpple.spot.server.common.jwt.UserDetailsCustom;
+import org.com.itpple.spot.server.common.auth.userDetails.UserDetailsCustom;
 import org.com.itpple.spot.server.dto.pot.request.CreatePotRequest;
 import org.com.itpple.spot.server.dto.pot.request.UploadImageRequest;
 import org.com.itpple.spot.server.dto.pot.response.CreatePotResponse;
@@ -53,9 +53,9 @@ public class PotController {
 
     @PostMapping(value = "", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<CreatePotResponse> createPot(
-            @AuthenticationPrincipal UserDetailsCustom userDetails,
+            @AuthenticationPrincipal UserDetailsCustom userDetailsCustom,
             @Valid @RequestBody CreatePotRequest createPotRequest) {
-        var userId = userDetails.getUser().getId();
+        var userId = userDetailsCustom.getUser().getId();
         return ResponseEntity.ok(potService.createPot(userId, createPotRequest));
     }
 }
